requirements.txt
streamlit
requests
Pillow
streamlit_cropper